plugins/noptilus/TransponderEstimation: Fix unlikely argument type problem.

+++ plugins-dev/noptilus/pt/lsts/neptus/plugins/noptilus/TransponderEstimation.java
@@ -218,7 +218,7 @@ public class TransponderEstimation extends ConsolePanel implements Renderer2DPai
 
         Random rand = new Random(System.currentTimeMillis());
 
-        for (long beacon : beaconNames.keySet()) {
+        for (short beacon : beaconNames.keySet()) {
             double[] estimate = Arrays.copyOf(beaconLocations.get(beacon), beaconLocations.get(beacon).length);
             double[] bestEstimate = Arrays.copyOf(estimate, estimate.length);
             
@@ -234,7 +234,6 @@ public class TransponderEstimation extends ConsolePanel implements Renderer2DPai
                 double r = 0;
 
                 for (int i = 0; i < measurements.size(); i++) {
-
                     double px = measurements.get(i)[X];
                     double py = measurements.get(i)[Y];
                     double pz = measurements.get(i)[Z];
